[Network] Rewrote network buffering; Removed network message packet features

+++ src/network/win32/winsockwrappers.cpp
@@ -329,7 +329,7 @@ ERR win_listen(WSW_SOCKET SocketHandle, int BackLog)
 
 //********************************************************************************************************************
 
-ERR WIN_RECEIVE(WSW_SOCKET SocketHandle, void *Buffer, int Len, int Flags, int *Result)
+ERR WIN_RECEIVE(WSW_SOCKET SocketHandle, void *Buffer, size_t Len, int Flags, int *Result)
 {
    *Result = 0;
    if (!Len) return ERR::Okay;
@@ -347,7 +347,7 @@ ERR WIN_RECEIVE(WSW_SOCKET SocketHandle, void *Buffer, int Len, int Flags, int *
 
 //********************************************************************************************************************
 
-ERR WIN_SEND(WSW_SOCKET Socket, const void *Buffer, int *Length, int Flags)
+ERR WIN_SEND(WSW_SOCKET Socket, const void *Buffer, size_t *Length, int Flags)
 {
    if (!*Length) return ERR::Okay;
    *Length = send(Socket, reinterpret_cast<const char *>(Buffer), *Length, Flags);
